Create build-bundle-with-qlpacks.yml
The Customizations.qll is a ql library that is included in every language library (e.g., java.qll ) which is the first thing one includes in a query  using for example import java . The action is copying user libraries with customizations into a folder customizations and adds an import of those libraries into Customizations.qll .

What changed with packaging is that when you build a query pack it includes all the dependencies to make it self contained. That is why it reused the library package (e.g., java-all) already part of its dependencies in .codeql/libraries residing in the query pack (e.q., java-queries) instead of the one we modified (java-all).

What the change is doing is rebuilding the query pack and replacing the contents of query pack in the CodeQL bundle with result of that rebuild. During the rebuild we are pointing to our modified library pack. The rebuild pulls in the dependencies and rebuilds the query cache.